[update] 投稿フォームをコンポーネント化し、不要なコードを整理
投稿フォームを `PostForm` コンポーネントとして分離し、重複するコードや不要なデータ定義を削除しました。これによりコードの可読性と再利用性を向上させました。
